+++ app/src/main/java/ru/surf/learn2invest/noui/database_components/entity/AssetBalanceHistory.kt
@@ -8,5 +8,5 @@ import androidx.room.PrimaryKey
 data class AssetBalanceHistory( // баланс портфеля
 
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val assetBalance: Double // стоимость портфеля
+    val assetBalance: Float // стоимость портфеля
 )
